requirements.txt
# データ収集
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.44.0

# データ処理
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30

# 機械学習
lightgbm==4.3.0
scikit-learn==1.4.2
shap==0.45.0

# UI
streamlit==1.35.0
plotly==5.22.0

# ユーティリティ
schedule==1.2.2
loguru==0.7.2
python-dotenv==1.0.1
